@@ -12,58 +12,67 @@
     "@ethersproject/experimental": "^5.6.0",
     "@ethersproject/providers": "^5.6.0",
     "@ethersproject/units": "^5.6.0",
-    "@fortawesome/fontawesome-svg-core": "^1.2.35",
-    "@fortawesome/free-brands-svg-icons": "^5.15.3",
-    "@fortawesome/free-solid-svg-icons": "^5.15.3",
+    "@fortawesome/fontawesome-svg-core": "^6.1.1",
+    "@fortawesome/free-brands-svg-icons": "^6.1.1",
+    "@fortawesome/free-solid-svg-icons": "^6.1.1",
     "@fortawesome/react-fontawesome": "^0.1.14",
-    "@reach/portal": "^0.16.0",
-    "@types/jest": "^26.0.9",
+    "@reach/portal": "^0.17.0",
+    "@types/jest": "^27.5.1",
     "@types/js-cookie": "^3.0.1",
-    "@types/node": "^14.0.27",
-    "@types/react": "^16.9.46",
-    "@types/react-tabs": "^2.3.2",
-    "@visx/responsive": "^1.10.1",
+    "@types/node": "^17.0.35",
+    "@types/react": "^18.0.9",
+    "@visx/curve": "^2.1.0",
+    "@visx/responsive": "^2.10.0",
     "@visx/scale": "^2.1.0",
-    "@visx/shape": "^1.17.1",
-    "@visx/visx": "^1.18.1",
+    "@visx/shape": "^2.10.0",
+    "@visx/visx": "^2.10.0",
     "@web3-react/abstract-connector": "^6.0.7",
     "@web3-react/core": "clabby/@web3-react/core",
     "@web3-react/injected-connector": "^6.0.7",
     "@web3-react/network-connector": "^6.2.9",
     "@web3-react/walletconnect-connector": "^6.2.13",
     "@web3-react/walletlink-connector": "^6.2.13",
+    "assert": "^2.0.0",
     "bignumber.js": "^9.0.1",
-    "bootstrap": "^5.0.2",
+    "bootstrap": "^5.2.0-beta1",
+    "buffer": "^6.0.3",
+    "crypto-browserify": "^3.12.0",
     "d3-array": "^3.0.1",
     "ethereum-multicall": "^2.7.1",
-    "ethers": "^5.0.2",
-    "graphql": "^15.5.1",
+    "ethers": "^5.6.8",
+    "graphql": "^16.5.0",
+    "https-browserify": "^1.0.0",
     "js-cookie": "^3.0.1",
     "lodash": "^4.17.21",
+    "os-browserify": "^0.3.0",
     "polished": "^4.1.3",
-    "react": "^16.13.1",
-    "react-bootstrap": "^2.0.0-beta.4",
-    "react-circular-progressbar": "^2.0.3",
+    "process": "^0.11.10",
+    "react": "^18.1.0",
+    "react-bootstrap": "^2.4.0",
+    "react-circular-progressbar": "^2.0.4",
     "react-countdown": "^2.2.1",
-    "react-countup": "^4.3.3",
-    "react-dom": "^16.13.1",
+    "react-countup": "^6.2.0",
+    "react-dom": "^18.1.0",
     "react-feather": "^2.0.9",
     "react-hook-mathjax": "^2.0.5",
     "react-lottie": "^1.2.3",
     "react-popper": "^2.2.5",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.1",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "5.0.1",
     "react-spring": "^9.2.4",
-    "react-tabs": "^3.1.2",
+    "react-tabs": "^5.1.0",
+    "stream-browserify": "^3.0.0",
+    "stream-http": "^3.2.0",
     "styled-components": "^5.1.1",
-    "typescript": "^3.9.7",
-    "web3": "^1.5.1",
-    "web3-utils": "^1.5.1"
+    "typescript": "^4.5.0",
+    "url": "^0.11.0",
+    "web3": "^1.7.3",
+    "web3-utils": "^1.7.3"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "CI= GENERATE_SOURCEMAP=false react-scripts build",
-    "test": "react-scripts test",
+    "start": "react-app-rewired start",
+    "build": "react-app-rewired build",
+    "test": "react-app-rewired test",
     "eject": "react-scripts eject",
     "deploy": "yarn build && cp ./build/index.html ./build/200.html",
     "lint": "npx eslint .",
