@@ -1,23 +1,8 @@
-import { Web3ReactProvider, createWeb3ReactRoot } from '@web3-react/core'
 import React from 'react'
-import ReactDOM from 'react-dom'
-import Web3 from 'web3'
-import { provider } from 'web3-core'
 import App from './App'
+import { createRoot } from 'react-dom/client'
 
-function getLibrary(provider: provider) {
-	return new Web3(provider)
-}
+const container = document.getElementById('root')
 
-const Web3ReactNetworkProvider = createWeb3ReactRoot('network')
-
-ReactDOM.render(
-	<React.StrictMode>
-		<Web3ReactProvider getLibrary={getLibrary}>
-			<Web3ReactNetworkProvider getLibrary={getLibrary}>
-				<App />
-			</Web3ReactNetworkProvider>
-		</Web3ReactProvider>
-	</React.StrictMode>,
-	document.getElementById('root'),
-)
+const root = createRoot(container)
+root.render(<App />)
