@@ -1,2 +1,1 @@
-export { default } from './Footer';
-
+export { default } from './Footer'
