@@ -1,2 +1,1 @@
-export { default } from './IconButton';
-
+export { default } from './IconButton'
