@@ -1,9 +1,13 @@
-import overlay from 'assets/img/overlay.png'
 import React from 'react'
+import { PropsWithChildren } from 'react'
 import styled from 'styled-components'
 import Footer from '../Footer'
 
-const Page: React.FC = ({ children }) => (
+interface PageProps {
+	children: any
+}
+
+const Page: React.FC<PropsWithChildren<PageProps>> = ({ children }) => (
 	<StyledPageContainer>
 		<StyledPageWrapper>
 			<StyledMain>{children}</StyledMain>
@@ -18,14 +22,13 @@ const StyledPageContainer = styled.div`
 	top: ${(props) => props.theme.topBarSize}px;
 	left: 0;
 	height: calc(100vh - ${(props) => props.theme.topBarSize}px);
-	width: 100%;
+	width: 100%; ;
 `
 
 const StyledPageWrapper = styled.div`
 	display: table-cell;
 	vertical-align: middle;
 	min-height: calc(100vh - ${(props) => props.theme.topBarSize}px);
-	background-image: url(${overlay});
 	background-size: cover;
 	background-repeat: no-repeat;
 	top: 0;
