@@ -2,11 +2,11 @@ import React from 'react'
 import styled from 'styled-components'
 
 export interface ProgressBarProps {
-    assetColor: string
-    width: number
-    label?: string
-  }
-  
+	assetColor: string
+	width: number
+	label?: string
+}
+
 export const Progress: React.FC<ProgressBarProps> = ({
 	assetColor,
 	width,
@@ -29,15 +29,15 @@ export const Progress: React.FC<ProgressBarProps> = ({
 }
 
 export const OutsideLabel = styled.span`
-float: left;
-margin-left: ${(props) => props.theme.spacing[2]}px;
-`  
+	float: left;
+	margin-left: ${(props) => props.theme.spacing[2]}px;
+`
 
 export const ProgressBar = styled.div.attrs((props: ProgressBarProps) => ({}))`
-  background-color: ${(props: ProgressBarProps) => props.assetColor};
-  border-radius: 5px;
-  width: ${(props: ProgressBarProps) => props.width}%;
-  height: 100%;
-  display: inline-block;
-  float: left;
+	background-color: ${(props: ProgressBarProps) => props.assetColor};
+	border-radius: 5px;
+	width: ${(props: ProgressBarProps) => props.width}%;
+	height: 100%;
+	display: inline-block;
+	float: left;
 `
