@@ -17,8 +17,8 @@ const useStakedBalance = (pid: number) => {
   const fetchBalance = useCallback(async () => {
     BigNumber.config({ DECIMAL_PLACES: 18 })
     const balance = await getStaked(masterChefContract, pid, account)
-    let userBalance
-    userBalance = new BigNumber(balance)
+    const userBalance = new BigNumber(balance)
+
     setBalance(userBalance.decimalPlaces(18))
   }, [account, pid, masterChefContract])
 
