@@ -26,7 +26,6 @@ import TokenInput from 'components/TokenInput'
 import { PoolType } from 'contexts/Farms/types'
 import { ethers } from 'ethers'
 import useAllowance from 'hooks/base/useAllowance'
-import useApprove from 'hooks/base/useApprove'
 import useBao from 'hooks/base/useBao'
 import useBlockDiff from 'hooks/base/useBlockDiff'
 import useTransactionHandler from 'hooks/base/useTransactionHandler'
@@ -35,7 +34,7 @@ import useFees from 'hooks/farms/useFees'
 import useStakedBalance from 'hooks/farms/useStakedBalance'
 import { useUserFarmInfo } from 'hooks/farms/useUserFarmInfo'
 import { default as React, useCallback, useMemo, useState } from 'react'
-import { Col, Modal, Row, Spinner } from 'react-bootstrap'
+import { Col, Modal, Row } from 'react-bootstrap'
 import {
 	exponentiate,
 	getDisplayBalance,
@@ -60,12 +59,10 @@ interface StakeProps {
 
 export const Stake: React.FC<StakeProps> = ({
 	lpContract,
-	lpTokenAddress,
 	pid,
 	poolType,
 	max,
 	tokenName = '',
-	ref = '',
 	pairUrl = '',
 	onHide,
 }) => {
@@ -183,6 +180,7 @@ export const Stake: React.FC<StakeProps> = ({
 												<a
 													href={`${Config.defaultRpc.blockExplorerUrls}/tx/${pendingTx}`}
 													target="_blank"
+													rel="noreferrer"
 												>
 													Pending Transaction{' '}
 													<FontAwesomeIcon icon="external-link-alt" />
@@ -200,9 +198,7 @@ export const Stake: React.FC<StakeProps> = ({
 												parseFloat(val) > max.toNumber()
 											}
 											onClick={async () => {
-												let stakeTx
-
-												stakeTx = masterChefContract.methods
+												const stakeTx = masterChefContract.methods
 													.deposit(
 														pid,
 														ethers.utils.parseUnits(val.toString(), 18),
@@ -225,9 +221,7 @@ export const Stake: React.FC<StakeProps> = ({
 								<SubmitButton
 									disabled={true}
 									onClick={async () => {
-										let stakeTx
-
-										stakeTx = masterChefContract.methods
+										const stakeTx = masterChefContract.methods
 											.deposit(
 												pid,
 												ethers.utils.parseUnits(val.toString(), 18),
@@ -264,13 +258,10 @@ interface UnstakeProps {
 }
 
 export const Unstake: React.FC<UnstakeProps> = ({
-	farm,
 	max,
 	tokenName = '',
 	pid = null,
-	ref = '',
 	pairUrl = '',
-	lpTokenAddress = '',
 	onHide,
 }) => {
 	const bao = useBao()
@@ -378,6 +369,7 @@ export const Unstake: React.FC<UnstakeProps> = ({
 									<a
 										href={`${Config.defaultRpc.blockExplorerUrls}/tx/${pendingTx}`}
 										target="_blank"
+										rel="noreferrer"
 									>
 										Pending Transaction{' '}
 										<FontAwesomeIcon icon="external-link-alt" />
@@ -396,14 +388,12 @@ export const Unstake: React.FC<UnstakeProps> = ({
 									stakedBalance.eq(new BigNumber(0))
 								}
 								onClick={async () => {
-									let unstakeTx
-
 									const amount =
 										val && isNaN(val as any)
 											? exponentiate(val, 18)
 											: new BigNumber(0).toFixed(4)
 
-									unstakeTx = masterChefContract.methods
+									const unstakeTx = masterChefContract.methods
 										.withdraw(
 											pid,
 											ethers.utils.parseUnits(val, 18),
@@ -466,6 +456,7 @@ export const Rewards: React.FC<RewardsProps> = ({ pid }) => {
 									<a
 										href={`${Config.defaultRpc.blockExplorerUrls}/tx/${pendingTx}`}
 										target="_blank"
+										rel="noreferrer"
 									>
 										Pending Transaction{' '}
 										<FontAwesomeIcon icon="external-link-alt" />
@@ -478,9 +469,7 @@ export const Rewards: React.FC<RewardsProps> = ({ pid }) => {
 							<SubmitButton
 								disabled={!earnings.toNumber()}
 								onClick={async () => {
-									let harvestTx
-
-									harvestTx = masterChefContract.methods
+									const harvestTx = masterChefContract.methods
 										.claimReward(pid)
 										.send({ from: account })
 
