@@ -1,2 +1,1 @@
-export { default } from './Farms';
-
+export { default } from './Farms'
